Add view function logic to post.py
TODO: Find an way to avoid copy-paste for finding authorized_posts

+++ hjlog/views/post.py
@@ -1,7 +1,7 @@
 from hjlog import db
 
 from flask import render_template, redirect, request, url_for, flash
-from sqlalchemy import desc
+from sqlalchemy import desc, or_
 from hjlog.models import Post, Photo, Tag
 from hjlog.forms import PostForm
 from flask.ext.login import current_user, login_required
@@ -10,6 +10,7 @@ import os
 msg_invalid_category = '존재하지 않는 카테고리입니다!'
 msg_invalid_input = '이런, 뭔가 빼먹으신 모양인데요?'
 msg_delete_success = '성공적으로 삭제되었습니다 :)'
+msg_unauthorized = '이 글에 접근할 권한이 없습니다.'
 
 def register(app):
     @app.route('/posts/<category>', defaults={'page': 1})
@@ -19,8 +20,17 @@ def register(app):
             flash(msg_invalid_category, 'warning')
             return redirect(url_for('posts', category='everyday'))
 
-        pgn = Post.query.filter_by(category=category).order_by(desc(Post.datetime))\
-            .paginate(int(page), per_page=10)
+        if current_user and current_user.is_authenticated():
+            accessible_posts = Post.query.filter(
+                    or_(~Post.private, Post.author == current_user))
+        else:
+            accessible_posts = Post.query.filter_by(private=False)
+
+        pgn = accessible_posts\
+                .filter_by(category=category)\
+                .order_by(desc(Post.datetime))\
+                .paginate(int(page), per_page=10)
+
         return render_template('posts.html', posts=pgn.items, c=category, pgn=pgn)
 
     @app.route('/post/<id>')
@@ -28,7 +38,17 @@ def register(app):
         post = Post.query.filter_by(id=id).one()
         category = post.category
 
-        base = Post.query.filter_by(category=category)
+        if post.private and post.author != current_user:
+            flash(msg_unauthorized, 'warning')
+            return redirect(url_for('about'))
+
+        if current_user and current_user.is_authenticated():
+            accessible_posts = Post.query.filter(
+                    or_(~Post.private, Post.author == current_user))
+        else:
+            accessible_posts = Post.query.filter_by(private=False)
+
+        base = accessible_posts.filter_by(category=category)
 
         prv = base.order_by(desc("id")).filter(Post.id<id).first()
         nxt = base.order_by("id").filter(Post.id>id).first()
@@ -49,14 +69,14 @@ def register(app):
 
         # Form Data Submitted
         if form.validate_on_submit():
-            # Valid input
-            title, body, category, author, tag_names = (
-                request.form.get('title'), request.form.get('body'),
-                category, current_user,
-                request.form.get('tags'))
+            keys = ['title', 'body', 'private', 'tags']
+            title, body, private, tag_names =\
+                    [request.form.get(key) for key in keys]
+
+            private = bool(private)
             tags = create_tags(tag_names)
 
-            post = Post(title, body, category, author, tags)
+            post = Post(title, body, category, current_user, private, tags)
             db.session.add(post)
             db.session.commit()
 
@@ -74,9 +94,13 @@ def register(app):
     def post_edit(id):
         post = Post.query.filter_by(id=id).one()
 
+        if post.private and post.author != current_user:
+            flash(msg_unauthorized, 'warning')
+            return redirect(url_for('about'))
+
         tags_str = ", ".join([tag.tag_name for tag in post.tags])
-        form = PostForm(title=post.title, body=post.body, tags=tags_str,
-                        category=post.category)
+        form = PostForm(title=post.title, body=post.body, private=post.private,
+                tags=tags_str, category=post.category)
 
         if request.method == 'GET':
             return render_template('post_edit.html', form=form, id=post.id)
@@ -85,13 +109,15 @@ def register(app):
         if form.validate_on_submit():
             old_tags = post.tags
 
-            # Valid input
-            title, body, category, tag_names = (
-                request.form.get('title'), request.form.get('body'),
-                request.form.get('category'), request.form.get('tags'))
+            keys = ['title', 'body', 'category', 'private', 'tags']
+            title, body, category, private, tag_names =\
+                    [request.form.get(key) for key in keys]
+
+            private = bool(private)
             tags = create_tags(tag_names)
 
-            post.title, post.body, post.category, post.tags = title, body, category, tags
+            post.title, post.body, post.category, post.private, post.tags =\
+                    title, body, category, private, tags
             db.session.add(post)
             db.session.commit()
 
@@ -111,6 +137,10 @@ def register(app):
     def post_delete(id):
         post = Post.query.filter_by(id=id).one()
 
+        if post.private and post.author != current_user:
+            flash(msg_unauthorized, 'warning')
+            return redirect(url_for('about'))
+
         photos = Photo.query.filter_by(original_id=id).all()
         delete_photos(photos, app.config['UPLOAD_FOLDER'])
 
@@ -127,8 +157,15 @@ def register(app):
     @app.route('/search/<tag_name>', defaults={'page': 1})
     @app.route('/search/<tag_name>/<page>')
     def search(tag_name, page):
-        pgn = Post.query.filter(Post.tags.any(tag_name=tag_name))\
-            .paginate(int(page), per_page=10)
+        if current_user and current_user.is_authenticated():
+            accessible_posts = Post.query.filter(
+                    or_(~Post.private, Post.author == current_user))
+        else:
+            accessible_posts = Post.query.filter_by(private=False)
+
+        pgn = accessible_posts\
+                .filter(Post.tags.any(tag_name=tag_name))\
+                .paginate(int(page), per_page=10)
 
         return render_template('search.html', posts=pgn.items, pgn=pgn,
                 tag_name=tag_name)
